Created First Simple View and asset
It's import to note you need `import play.api._` to use the views object
when selecting the template for a controller response.

Another thing to note is `@routes.Assets.at` to retrieve the URL for
your assets, in order for this to compile you need to have the following
in your routes file:

	GET /assets/*file controllers.Assets.at(path="/public", file)

These are things which the documentation doesn't make perfectly clear,
it might be good to write up a beginners blog post about this I think.